Internal #861: Refactor Window Ranks
Move the window Rank function classes to separate files.
